server/requirements.txt
flask
python-dotenv
pymysql
numpy
sentence-transformers
langchain
langchain-community
langchain-text-splitters
langchain-chroma
langchain-huggingface
transformers
torch


pip install flask python-dotenv mysql-connector-python requests faiss-cpu numpy